+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequestStartLine.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -7,8 +7,8 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class RequestStartLine {
-    private static final Logger log = LoggerFactory.getLogger(RequestStartLine.class);
+public class HttpRequestStartLine {
+    private static final Logger log = LoggerFactory.getLogger(HttpRequestStartLine.class);
     private static final int INDEX_REQUEST_URI = 1;
     private static final int INDEX_HTTP_METHOD = 0;
     private static final int INDEX_HTTP_VERSION = 2;
@@ -18,17 +18,17 @@ public class RequestStartLine {
     private final String requestURI;
     private final String httpVersion;
 
-    public RequestStartLine(HttpMethod httpMethod, String requestURI, String httpVersion) {
+    public HttpRequestStartLine(HttpMethod httpMethod, String requestURI, String httpVersion) {
         this.httpMethod = httpMethod;
         this.requestURI = requestURI;
         this.httpVersion = httpVersion;
     }
 
-    public static RequestStartLine from(String request) {
+    public static HttpRequestStartLine from(String request) {
         List<String> requests = Arrays.stream(request.split(" ")).collect(Collectors.toList());
         validateRequestStartLineSize(request, requests);
 
-        return new RequestStartLine(
+        return new HttpRequestStartLine(
                 HttpMethod.of(requests.get(INDEX_HTTP_METHOD)),
                 requests.get(INDEX_REQUEST_URI),
                 requests.get(INDEX_HTTP_VERSION)
